requirements.txt
streamlit
pytesseract
Pillow
whisper
PyMuPDF
torch